Time: 13 ms (89.18%), Space: 78 MB (56.86%) - LeetHub

+++ 1477-product-of-the-last-k-numbers/1477-product-of-the-last-k-numbers.cpp
@@ -1,6 +1,6 @@
 class ProductOfNumbers {
 public:
-    vector<long long> mul;
+    vector<int> mul;
     int lastZeroPos = -1;
     int len = 0;
     void add(int num) {
